Fixed frontend request for authenticated user
