+++ backend/config/keys.js
@@ -1,11 +1,16 @@
-const KEYS = {
+const GOOGLE_TOKENS = {
     GOOGLE_CLIENT_ID:
         "418890674873-o2b40alq0cudvp4dc60lgfr5ii7tfv6h.apps.googleusercontent.com",
     GOOGLE_CLIENT_SECRET: "-Ib0-fPZl15s__R9KkxuBl6T",
 };
 
 const SESSION = {
-    COOKIE_KEY: "thisappisawesome",
+    COOKIE_KEY: "3DCreative",
+};
+
+const KEYS = {
+    ...GOOGLE_TOKENS,
+    ...SESSION,
 };
 
 module.exports = KEYS;
