+++ backend/routes/auth.js
@@ -42,24 +42,4 @@ router.get(
     })
 );
 
-// const authCheck = (req, res, next) => {
-//     if (!req.user) {
-//         res.status(401).json({
-//             authenticated: false,
-//             message: "User has not been authenticated.",
-//         });
-//     } else {
-//         next();
-//     }
-// };
-
-// router.get("/", authCheck, (req, res) => {
-//     res.status(200).json({
-//         authenticated: true,
-//         message: "User successfully authenticated.",
-//         user: req.user,
-//         cookies: req.cookies,
-//     });
-// });
-
 module.exports = router;
