+++ backend/server.js
@@ -4,8 +4,6 @@ const cookieSession = require("cookie-session");
 const passport = require("passport");
 const passportSetup = require("./config/passport-setup");
 const session = require("express-session");
-const authRoutes = require("./routes/auth");
-const mongoose = require("mongoose");
 const keys = require("./config/keys");
 const cookieParser = require("cookie-parser");
 const express = require("express");
