+++ frontend/src/components/static/auth/LogoutButton.js
@@ -1,16 +1,11 @@
 import React from "react";
-import { GoogleLogout } from "react-google-login";
+import Button from "@material-ui/core/Button";
 
-const LogoutButton = ({ userHasAuthenticated }) => {
-    const logout = userHasAuthenticated(false);
-
-    return (
-        <GoogleLogout
-            clientId="418890674873-o2b40alq0cudvp4dc60lgfr5ii7tfv6h.apps.googleusercontent.com"
-            buttonText="Logout"
-            onLogoutSuccess={logout}
-        />
-    );
+const LogoutButton = () => {
+    const handleLogoutButton = () => {
+        window.open("http://localhost:5000/auth/logout", "_self");
+    };
+    return <Button onClick={handleLogoutButton}>Atsijungti</Button>;
 };
 
 export default LogoutButton;
